teacher form
When a teacher activates their account, they get redirected to an empty calendar. This calendar has two views, weekly and monthly, For weekly, they can either just click, or click and drag their available time slots. for monthly, likely to yousefs calendar thing, it adds a popup where you can add a note, and your start time and end time, as well as that thing on the bottom, that might not be useful we'll see. this also adds to the database and it is functional